Bug/sonar qube duplication fix (#119)
* Seperating into 2 controllers - Registration and registrationmaterial

* Sonal Qube duplicate issue

* Applied API Convention attribute

* Fixing sonar duplicate issue on action metadata

* excluded static class from code coverage

---------
